Split certificates into multiple files
